+++ hibernate6/src/main/java/com/github/starnowski/posjsonhelper/hibernate6/descriptor/AbstractJsonBExtractPathDescriptorRegister.java
@@ -21,44 +21,34 @@
  */
 package com.github.starnowski.posjsonhelper.hibernate6.descriptor;
 
+import com.github.starnowski.posjsonhelper.hibernate6.AbstractJsonBExtractPath;
 import org.hibernate.query.sqm.function.SqmFunctionDescriptor;
 import org.hibernate.query.sqm.function.SqmFunctionRegistry;
 
 /**
- * Type extends {@link AbstractConditionalFunctionDescriptorRegister} type.
- * It registers function with key {@link #hqlFunctionName}
- * and link it with SQL function with name {@link #sqlFunctionName}.
+ * Type extends {@link  AbstractConditionalFunctionDescriptorRegister} type.
+ * Responsible for register of hql function for any child type of {@link  AbstractJsonBExtractPath}.
+ * It uses component of type {@link AbstractJsonBExtractPathDescriptor} for rendering.
  */
-public class FunctionByNameRegister extends AbstractConditionalFunctionDescriptorRegister {
+public class AbstractJsonBExtractPathDescriptorRegister extends AbstractConditionalFunctionDescriptorRegister {
+
+    private final AbstractJsonBExtractPathDescriptor abstractJsonBExtractPathDescriptor;
 
     /**
-     * Key on which the hql function is going to be registered
+     * @param shouldOverrideFunctionIfAlreadyRegistered value of property {@link #shouldOverrideFunctionIfAlreadyRegistered}
      */
-    private final String hqlFunctionName;
-    /**
-     * Name of sql function
-     */
-    private final String sqlFunctionName;
-
-    /**
-     *
-     * @param hqlFunctionName value of {@link #hqlFunctionName}
-     * @param sqlFunctionName value of {@link #sqlFunctionName}
-     * @param shouldOverrideFunctionIfAlreadyRegistered value of {{@link #shouldOverrideFunctionIfAlreadyRegistered}}
-     */
-    public FunctionByNameRegister(String hqlFunctionName, String sqlFunctionName, boolean shouldOverrideFunctionIfAlreadyRegistered) {
+    public AbstractJsonBExtractPathDescriptorRegister(AbstractJsonBExtractPathDescriptor abstractJsonBExtractPathDescriptor, boolean shouldOverrideFunctionIfAlreadyRegistered) {
         super(shouldOverrideFunctionIfAlreadyRegistered);
-        this.hqlFunctionName = hqlFunctionName;
-        this.sqlFunctionName = sqlFunctionName;
+        this.abstractJsonBExtractPathDescriptor = abstractJsonBExtractPathDescriptor;
     }
 
     @Override
     protected SqmFunctionDescriptor register(SqmFunctionRegistry registry) {
-        return registry.namedDescriptorBuilder(hqlFunctionName, sqlFunctionName).register();
+        return registry.register(abstractJsonBExtractPathDescriptor.getName(), abstractJsonBExtractPathDescriptor);
     }
 
     @Override
     protected String getHqlFunctionName() {
-        return hqlFunctionName;
+        return abstractJsonBExtractPathDescriptor.getName();
     }
 }