+++ hibernate6/src/main/java/com/github/starnowski/posjsonhelper/hibernate6/descriptor/JsonbAnyArrayStringsExistPredicateDescriptor.java
@@ -19,28 +19,31 @@
  *     Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
  *     USA
  */
-package com.github.starnowski.posjsonhelper.hibernate6.predicates;
+package com.github.starnowski.posjsonhelper.hibernate6.descriptor;
 
+import com.github.starnowski.posjsonhelper.core.Context;
 import com.github.starnowski.posjsonhelper.core.HibernateContext;
 import com.github.starnowski.posjsonhelper.hibernate6.JsonBExtractPath;
 import com.github.starnowski.posjsonhelper.hibernate6.operators.JsonArrayFunction;
+import com.github.starnowski.posjsonhelper.hibernate6.predicates.JsonbAnyArrayStringsExistPredicate;
 import org.hibernate.query.sqm.NodeBuilder;
 
 /**
- * Type that extends {@link AbstractJsonbArrayStringsExistPredicate}.
- * Implemented of HQL function defined by method {@link HibernateContext#getJsonbAnyArrayStringsExistOperator()} }
+ * Function descriptor for child type of {@link JsonbAnyArrayStringsExistPredicate}
  */
-public class JsonbAnyArrayStringsExistPredicate extends AbstractJsonbArrayStringsExistPredicate<JsonbAnyArrayStringsExistPredicate> {
-    public JsonbAnyArrayStringsExistPredicate(HibernateContext context, NodeBuilder nodeBuilder, JsonBExtractPath jsonBExtractPath, String[] values) {
-        super(context, nodeBuilder, jsonBExtractPath, values, context.getJsonbAnyArrayStringsExistOperator());
-    }
-
-    public JsonbAnyArrayStringsExistPredicate(HibernateContext context, NodeBuilder nodeBuilder, JsonBExtractPath jsonBExtractPath, JsonArrayFunction jsonArrayFunction) {
-        super(context, nodeBuilder, jsonBExtractPath, jsonArrayFunction, context.getJsonbAnyArrayStringsExistOperator());
+public class JsonbAnyArrayStringsExistPredicateDescriptor extends AbstractJsonbArrayStringsExistPredicateDescriptor<JsonbAnyArrayStringsExistPredicate> {
+    public JsonbAnyArrayStringsExistPredicateDescriptor(Context context, HibernateContext hibernateContext) {
+        super(context.getJsonbAnyArrayStringsExistFunctionReference(), hibernateContext);
     }
 
     @Override
-    protected JsonbAnyArrayStringsExistPredicate generateCopy(HibernateContext context, NodeBuilder nodeBuilder, JsonBExtractPath jsonBExtractPath, JsonArrayFunction jsonArrayFunction) {
-        return new JsonbAnyArrayStringsExistPredicate(context, nodeBuilder, jsonBExtractPath, jsonArrayFunction);
+    protected JsonbAnyArrayStringsExistPredicate generateJsonbArrayStringsExistPredicate(HibernateContext context, NodeBuilder nodeBuilder, JsonBExtractPath jsonBExtractPath, JsonArrayFunction arrayFunction) {
+        return new JsonbAnyArrayStringsExistPredicate(context, nodeBuilder, jsonBExtractPath, arrayFunction);
     }
+
+    @Override
+    public String getSqmFunction() {
+        return this.hibernateContext.getJsonbAnyArrayStringsExistOperator();
+    }
+
 }